move RouteMap to its own mini-library
