+++ lib/routemaster.dart
@@ -4,3 +4,4 @@ export 'src/core.dart';
 export 'src/pages/guard.dart';
 export 'src/pages/transition_page.dart';
 export 'src/parser.dart';
+export 'src/route_map.dart';
